- Removed commented out functions - Fixed some type stability functions - Commented out some probably not needed functions - Added threads to parallel map - Added some currently unused structs, need to test - Fixed the update for the global u and interpolant

+++ src/Grid/grids.jl
@@ -87,8 +87,8 @@ struct Grid2D{TT,
     gridy   :: GT
     Δx      :: TT
     Δy      :: TT
-    nx      :: Integer
-    ny      :: Integer
+    nx      :: Int64
+    ny      :: Int64
 
     J       :: GT
     qx      :: GT
